requirements.txt
qrcode[pil]
Pillow
python-barcode
